change all parameter's type in cobale syscall handle functions from ui to ul

+++ kernel/cobalt/posix/io.h
@@ -50,12 +50,12 @@ COBALT_SYSCALL_DECL(recvmsg,
 		    (int fd, struct user_msghdr __user *umsg, int flags));
 
 COBALT_SYSCALL_DECL(recvmmsg,
-		    (int fd, struct mmsghdr __user *u_msgvec, unsigned int vlen,
-		     unsigned int flags, struct __user_old_timespec __user *u_timeout));
+		    (int fd, struct mmsghdr __user *u_msgvec, unsigned long vlen,
+		     unsigned long flags, struct __user_old_timespec __user *u_timeout));
 
 COBALT_SYSCALL_DECL(recvmmsg64,
-		    (int fd, struct mmsghdr __user *u_msgvec, unsigned int vlen,
-		     unsigned int flags,
+		    (int fd, struct mmsghdr __user *u_msgvec, unsigned long vlen,
+		     unsigned long flags,
 		     struct __kernel_timespec __user *u_timeout));
 
 COBALT_SYSCALL_DECL(sendmsg,
@@ -63,7 +63,7 @@ COBALT_SYSCALL_DECL(sendmsg,
 
 COBALT_SYSCALL_DECL(sendmmsg,
 		    (int fd, struct mmsghdr __user *u_msgvec,
-		     unsigned int vlen, unsigned int flags));
+		     unsigned long vlen, unsigned long flags));
 
 COBALT_SYSCALL_DECL(mmap,
 		    (int fd, struct _rtdm_mmap_request __user *u_rma,
